something weird is happening

+++ src/great_risks/mcts_agent_greedy.hh
@@ -14,7 +14,7 @@ namespace great_risks
         GreedyAgent opp_greedy;
         uint8_t opp_index;
         std::mt19937 rng;
-        tsl::robin_map<Field, double> rollout_cache;
+        tsl::robin_map<Field, float> rollout_cache;
 
     public:
         MCTSAgentGreedy(uint8_t index, uint8_t opp_index, uint32_t seed = 5489)
